requirements.txt
Django==1.6.5
django_compressor==1.4
django-vanilla-views==1.0.3
gevent==1.0.1
gunicorn==19.0.0
pygments==1.6

-e git+https://github.com/odeoncg/django-pygments.git#egg=django-pygments
